Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ LeetCode/Hard/1220-count-vowels-permutation/1220-count-vowels-permutation.cpp
@@ -1,5 +1,5 @@
 #define ll long long
-#define Modulo (int)(1e9+7)
+#define Modulo (ll)(1e9+7)
 class Solution {
 public:
     int countVowelPermutation(int n) {
